requirements.txt
streamlit
pymupdf
python-docx
chromadb
sentence-transformers
pydantic
psutil